requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pypdf==6.1.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
langchain==0.1.0
langchain-google-genai==0.0.8
google-generativeai==0.3.2
pydantic==2.5.2
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
